Set user store domain thread local in auto login flow
Downstream code expects the user store domain to be set in the thread local. This thread local usually gets populated during any user store operation
at the user store layer. In the auto login flow, we do not call any userstore operations hence the thread local will not be set.

As a solution, we set the thread local explicitly. The value for the user store domain is expected to be set in the cookie.